Normalize average passband gain to 0 dB
