+++ tests/dsp/fir/test_design_lowpass_fir.py
@@ -3,6 +3,8 @@ import pytest
 
 import sdr
 
+from .helper import verify_impulse_response
+
 
 def test_exceptions():
     with pytest.raises(TypeError):
@@ -63,9 +65,7 @@ def test_custom():
             0.000000000000000,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 def test_hamming():
@@ -110,9 +110,7 @@ def test_hamming():
             0.000000000000000,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 def test_hann():
@@ -157,9 +155,7 @@ def test_hann():
             0,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 def test_blackman():
@@ -204,9 +200,7 @@ def test_blackman():
             -0.000000000000000,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 def test_blackman_harris():
@@ -251,9 +245,7 @@ def test_blackman_harris():
             0.000000000000000,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 def test_chebyshev():
@@ -298,9 +290,7 @@ def test_chebyshev():
             0.000000000000000,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 # NOTE: Added extra Kaiser window tests to reverse engineer MATLAB's beta parameter
@@ -348,9 +338,7 @@ def test_kaiser30_0p1():
             -0.019837202437853,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 def test_kaiser_30_0p2():
@@ -395,9 +383,7 @@ def test_kaiser_30_0p2():
             0.000000000000000,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 def test_kaiser30_0p3():
@@ -442,9 +428,7 @@ def test_kaiser30_0p3():
             0.019631739746596,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
+    verify_impulse_response(h, h_truth)
 
 
 def test_kaiser_60_0p2():
@@ -519,22 +503,4 @@ def test_kaiser_60_0p2():
             -0.000000000000000,
         ]
     )
-    # debug_plot(h, h_truth)
-    assert h.shape == h_truth.shape
-    assert np.allclose(h, h_truth)
-
-
-def debug_plot(h, h_truth):
-    import matplotlib.pyplot as plt
-
-    plt.figure()
-    sdr.plot.time_domain(h_truth, marker=".", label="h_truth")
-    sdr.plot.time_domain(h, marker="x", label="h")
-    plt.legend()
-
-    plt.figure()
-    sdr.plot.magnitude_response(h_truth, label="h_truth")
-    sdr.plot.magnitude_response(h, label="h")
-    plt.legend()
-
-    plt.show()
+    verify_impulse_response(h, h_truth)
